+++ tests/test_parsers.py
@@ -17,11 +17,11 @@ class TestParsers(unittest.TestCase):
         parser = ControlCodeParser()
         parser.reset("\0\b\ra[", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "^@"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "^H"))
-        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "^M"))
-        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "["))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "^@"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "^H"))
+        self.assertEqual(next(tokens), (2, Parser.DISPLAY_TEXT, "^M"))
+        self.assertEqual(next(tokens), (3, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, "["))
 
     def test_asciimatics_parser(self):
         """
@@ -30,15 +30,15 @@ class TestParsers(unittest.TestCase):
         parser = AsciimaticsParser()
         parser.reset("a${1}b${2,1}c${3,2,4}de${7}", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.CHANGE_COLOURS, (1, 0, None)))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
-        self.assertEquals(next(tokens), (6, Parser.CHANGE_COLOURS, (2, 1, None)))
-        self.assertEquals(next(tokens), (6, Parser.DISPLAY_TEXT, "c"))
-        self.assertEquals(next(tokens), (13, Parser.CHANGE_COLOURS, (3, 2, 4)))
-        self.assertEquals(next(tokens), (13, Parser.DISPLAY_TEXT, "d"))
-        self.assertEquals(next(tokens), (22, Parser.DISPLAY_TEXT, "e"))
-        self.assertEquals(next(tokens), (23, Parser.CHANGE_COLOURS, (7, 0, None)))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.CHANGE_COLOURS, (1, 0, None)))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (6, Parser.CHANGE_COLOURS, (2, 1, None)))
+        self.assertEqual(next(tokens), (6, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (13, Parser.CHANGE_COLOURS, (3, 2, 4)))
+        self.assertEqual(next(tokens), (13, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (22, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (23, Parser.CHANGE_COLOURS, (7, 0, None)))
         with self.assertRaises(StopIteration):
             next(tokens)
 
@@ -51,37 +51,37 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Normal text
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
 
         # Unknown escape code
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
 
         # Reset
-        self.assertEquals(next(tokens), (7, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
-        self.assertEquals(next(tokens), (7, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (7, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
+        self.assertEqual(next(tokens), (7, Parser.DISPLAY_TEXT, "c"))
 
         # Bold
-        self.assertEquals(next(tokens), (12, Parser.CHANGE_COLOURS, (7, constants.A_BOLD, 0)))
-        self.assertEquals(next(tokens), (12, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (12, Parser.CHANGE_COLOURS, (7, constants.A_BOLD, 0)))
+        self.assertEqual(next(tokens), (12, Parser.DISPLAY_TEXT, "d"))
 
         # Normal
-        self.assertEquals(next(tokens), (17, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
-        self.assertEquals(next(tokens), (17, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (17, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
+        self.assertEqual(next(tokens), (17, Parser.DISPLAY_TEXT, "e"))
 
         # Inverse
-        self.assertEquals(next(tokens), (22, Parser.CHANGE_COLOURS, (7, constants.A_REVERSE, 0)))
-        self.assertEquals(next(tokens), (22, Parser.DISPLAY_TEXT, "f"))
+        self.assertEqual(next(tokens), (22, Parser.CHANGE_COLOURS, (7, constants.A_REVERSE, 0)))
+        self.assertEqual(next(tokens), (22, Parser.DISPLAY_TEXT, "f"))
 
         # Unset inverse
-        self.assertEquals(next(tokens), (27, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
-        self.assertEquals(next(tokens), (27, Parser.DISPLAY_TEXT, "g"))
+        self.assertEqual(next(tokens), (27, Parser.CHANGE_COLOURS, (7, constants.A_NORMAL, 0)))
+        self.assertEqual(next(tokens), (27, Parser.DISPLAY_TEXT, "g"))
 
         # Standard colours, using multiple parameters
-        self.assertEquals(next(tokens), (33, Parser.CHANGE_COLOURS, (constants.COLOUR_RED, constants.A_NORMAL, constants.COLOUR_GREEN)))
-        self.assertEquals(next(tokens), (33, Parser.DISPLAY_TEXT, "h"))
+        self.assertEqual(next(tokens), (33, Parser.CHANGE_COLOURS, (constants.COLOUR_RED, constants.A_NORMAL, constants.COLOUR_GREEN)))
+        self.assertEqual(next(tokens), (33, Parser.DISPLAY_TEXT, "h"))
 
         # Final escape sequence with no visible text is returned with no text.
-        self.assertEquals(next(tokens), (42, Parser.CHANGE_COLOURS, (constants.COLOUR_WHITE, constants.A_NORMAL, constants.COLOUR_BLACK)))
+        self.assertEqual(next(tokens), (42, Parser.CHANGE_COLOURS, (constants.COLOUR_WHITE, constants.A_NORMAL, constants.COLOUR_BLACK)))
 
         with self.assertRaises(StopIteration):
             next(tokens)
@@ -95,15 +95,15 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Normal text
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
 
         # Default foreground colour
-        self.assertEquals(next(tokens), (1, Parser.CHANGE_COLOURS, (constants.COLOUR_DEFAULT, None, None)))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (1, Parser.CHANGE_COLOURS, (constants.COLOUR_DEFAULT, None, None)))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
 
         # Default background colour
-        self.assertEquals(next(tokens), (7, Parser.CHANGE_COLOURS, (constants.COLOUR_DEFAULT, None, constants.COLOUR_DEFAULT)))
-        self.assertEquals(next(tokens), (7, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (7, Parser.CHANGE_COLOURS, (constants.COLOUR_DEFAULT, None, constants.COLOUR_DEFAULT)))
+        self.assertEqual(next(tokens), (7, Parser.DISPLAY_TEXT, "c"))
 
     def test_ansi_terminal_parser_palette(self):
         """
@@ -115,25 +115,25 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Bad colour scheme - ignore
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
 
         # Standard colour palette
-        self.assertEquals(next(tokens), (8, Parser.CHANGE_COLOURS, (17, None, None)))
-        self.assertEquals(next(tokens), (8, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (8, Parser.CHANGE_COLOURS, (17, None, None)))
+        self.assertEqual(next(tokens), (8, Parser.DISPLAY_TEXT, "b"))
 
         # RGB colour scheme - ignore
-        self.assertEquals(next(tokens), (19, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (19, Parser.DISPLAY_TEXT, "c"))
 
         # Standard colour palette
-        self.assertEquals(next(tokens), (33, Parser.CHANGE_COLOURS, (17, None, 54)))
-        self.assertEquals(next(tokens), (33, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (33, Parser.CHANGE_COLOURS, (17, None, 54)))
+        self.assertEqual(next(tokens), (33, Parser.DISPLAY_TEXT, "d"))
 
         # Unknown parameter
-        self.assertEquals(next(tokens), (44, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (44, Parser.DISPLAY_TEXT, "e"))
 
         # Intense colour palette
-        self.assertEquals(next(tokens), (51, Parser.CHANGE_COLOURS, (11, None, 54)))
-        self.assertEquals(next(tokens), (51, Parser.CHANGE_COLOURS, (11, None, 12)))
+        self.assertEqual(next(tokens), (51, Parser.CHANGE_COLOURS, (11, None, 54)))
+        self.assertEqual(next(tokens), (51, Parser.CHANGE_COLOURS, (11, None, 12)))
 
     def test_ansi_terminal_parser_cursor(self):
         """
@@ -144,39 +144,39 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Normal text...
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "a"))
 
         # Backspace and overwrite.
-        self.assertEquals(next(tokens), (2, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (2, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (2, Parser.DISPLAY_TEXT, "b"))
 
         # Carriage return and overwrite
-        self.assertEquals(next(tokens), (4, Parser.MOVE_ABSOLUTE, (0, None)))
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (4, Parser.MOVE_ABSOLUTE, (0, None)))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, "c"))
 
         # Move cursor forwards and append.
-        self.assertEquals(next(tokens), (6, Parser.MOVE_RELATIVE, (1, 0)))
-        self.assertEquals(next(tokens), (6, Parser.DISPLAY_TEXT, "d"))
-        self.assertEquals(next(tokens), (10, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (6, Parser.MOVE_RELATIVE, (1, 0)))
+        self.assertEqual(next(tokens), (6, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (10, Parser.DISPLAY_TEXT, "d"))
 
         # Move cursor backwards and overwrite.
-        self.assertEquals(next(tokens), (11, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (11, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (11, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (11, Parser.DISPLAY_TEXT, "e"))
 
         # Move cursor up and down.
-        self.assertEquals(next(tokens), (15, Parser.MOVE_RELATIVE, (0, -1)))
-        self.assertEquals(next(tokens), (15, Parser.MOVE_RELATIVE, (0, 1)))
+        self.assertEqual(next(tokens), (15, Parser.MOVE_RELATIVE, (0, -1)))
+        self.assertEqual(next(tokens), (15, Parser.MOVE_RELATIVE, (0, 1)))
 
         # Move cursor to location
-        self.assertEquals(next(tokens), (15, Parser.MOVE_ABSOLUTE, (1, 0)))
+        self.assertEqual(next(tokens), (15, Parser.MOVE_ABSOLUTE, (1, 0)))
 
         # Show/hide cursor
-        self.assertEquals(next(tokens), (15, Parser.SHOW_CURSOR, True))
-        self.assertEquals(next(tokens), (15, Parser.SHOW_CURSOR, False))
+        self.assertEqual(next(tokens), (15, Parser.SHOW_CURSOR, True))
+        self.assertEqual(next(tokens), (15, Parser.SHOW_CURSOR, False))
 
         # Trailing Carriage return
-        self.assertEquals(next(tokens), (15, Parser.MOVE_ABSOLUTE, (0, None)))
+        self.assertEqual(next(tokens), (15, Parser.MOVE_ABSOLUTE, (0, None)))
 
     def test_ansi_terminal_parser_delete(self):
         """
@@ -187,54 +187,54 @@ class TestParsers(unittest.TestCase):
         # Delete to end of line
         parser.reset("abcde\x08\x08\x08\x1B[K", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
-        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
-        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
-        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (5, Parser.DELETE_LINE, 0))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (5, Parser.DELETE_LINE, 0))
         with self.assertRaises(StopIteration):
             next(tokens)
 
         # Delete to start of line
         parser.reset("abcde\x1B[1K", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
-        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
-        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
-        self.assertEquals(next(tokens), (5, Parser.DELETE_LINE, 1))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (5, Parser.DELETE_LINE, 1))
         with self.assertRaises(StopIteration):
             next(tokens)
 
         # Delete line
         parser.reset("abcde\x1B[2K", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
-        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
-        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
-        self.assertEquals(next(tokens), (5, Parser.DELETE_LINE, 2))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (5, Parser.DELETE_LINE, 2))
         with self.assertRaises(StopIteration):
             next(tokens)
 
         # Delete char
         parser.reset("abcde\x08\x08\x08\x1B[P", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
-        self.assertEquals(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
-        self.assertEquals(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
-        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
-        self.assertEquals(next(tokens), (5, Parser.DELETE_CHARS, 1))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (2, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (3, Parser.DISPLAY_TEXT, "d"))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, "e"))
+        self.assertEqual(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (5, Parser.MOVE_RELATIVE, (-1, 0)))
+        self.assertEqual(next(tokens), (5, Parser.DELETE_CHARS, 1))
         with self.assertRaises(StopIteration):
             next(tokens)
 
@@ -247,14 +247,14 @@ class TestParsers(unittest.TestCase):
         tokens = parser.parse()
 
         # Ignore unknown escape and next letter
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
 
         # ANSI art uses control codes for special characters - check we just blank them.
-        self.assertEquals(next(tokens), (4, Parser.DISPLAY_TEXT, " "))
+        self.assertEqual(next(tokens), (4, Parser.DISPLAY_TEXT, " "))
 
         # Back to normal.
-        self.assertEquals(next(tokens), (5, Parser.DISPLAY_TEXT, "c"))
+        self.assertEqual(next(tokens), (5, Parser.DISPLAY_TEXT, "c"))
 
     def test_ansi_terminal_parser_tab(self):
         """
@@ -263,7 +263,7 @@ class TestParsers(unittest.TestCase):
         parser = AnsiTerminalParser()
         parser.reset("\x09", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.NEXT_TAB, None))
+        self.assertEqual(next(tokens), (0, Parser.NEXT_TAB, None))
 
     def test_ansi_terminal_parser_clear(self):
         """
@@ -272,7 +272,7 @@ class TestParsers(unittest.TestCase):
         parser = AnsiTerminalParser()
         parser.reset("\x1B[2J", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.CLEAR_SCREEN, None))
+        self.assertEqual(next(tokens), (0, Parser.CLEAR_SCREEN, None))
 
     def test_ansi_terminal_parser_os_cmd(self):
         """
@@ -281,8 +281,8 @@ class TestParsers(unittest.TestCase):
         parser = AnsiTerminalParser()
         parser.reset("a\x1B]do something;stuff:to^ignore\x07b", None)
         tokens = parser.parse()
-        self.assertEquals(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
-        self.assertEquals(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
+        self.assertEqual(next(tokens), (0, Parser.DISPLAY_TEXT, "a"))
+        self.assertEqual(next(tokens), (1, Parser.DISPLAY_TEXT, "b"))
 
     def test_ansi_terminal_parser_bell(self):
         """