+++ CHANGES.rst
@@ -5,6 +5,7 @@ LATEST
 ------
 - Added support for ColouredText titles in MultiColumnLIstBox.
 - Added gutter option to Layout.
+- Added speed option to Sprite.
 - Fixed bug where moving focus between Frames resulted in no current focus.
 - Fixed internal state of RadioButton values to be consistent with selection.